refactor(shutdown): extract shared session sweep utility + fix stale comments
- Extract duplicate session sweep logic from container.ts and index.ts
  into src/services/session-sweep.ts (sweepTmuxSessions). The two call
  sites diverged: container silently ignored failures while index.ts
  logged warnings. The utility takes an optional Logger and preserves
  both behaviors from a single source of truth.
- Fix stale JSDoc in worker-repository.ts: "PID check remains
  authoritative" → "Tmux session check is authoritative" (Phase 4
  replaced PID-based liveness with tmux session checks).
- Add missing step 6 "Sweep remaining tmux sessions" to container.ts
  dispose() @remarks and renumber subsequent steps (6→7 Close DB,
  7→8 ShutdownComplete, 8→9 Dispose EventBus, 9→10 Clear services).

Addresses: I1-session-sweep-dup, I5-stale-pid-comment, I7-stale-remarks
