+++ src/core/container.ts
@@ -3,6 +3,7 @@
  * Manages service lifecycle and dependencies
  */
 
+import { sweepTmuxSessions } from '../services/session-sweep.js';
 import { AutobeatError, ErrorCode } from './errors.js';
 import { err, ok, Result } from './result.js';
 import type { TmuxSessionManagerCorePort } from './tmux-types.js';
@@ -167,10 +168,11 @@ export class Container {
    * 3. Stop ScheduleExecutor - prevents timer firing against closed resources
    * 4. Stop translation proxy - lets in-flight requests complete before workers die
    * 5. Kill all workers (emit `WorkersTerminating`, then `killAll()`)
-   * 6. Close database (emit `DatabaseClosing`, then `close()`)
-   * 7. Emit `ShutdownComplete` event
-   * 8. Dispose EventBus (clears cleanup timers)
-   * 9. Clear all service registrations
+   * 6. Sweep remaining tmux sessions (belt-and-suspenders for sessions missed by killAll)
+   * 7. Close database (emit `DatabaseClosing`, then `close()`)
+   * 8. Emit `ShutdownComplete` event
+   * 9. Dispose EventBus (clears cleanup timers)
+   * 10. Clear all service registrations
    *
    * @example
    * ```typescript
@@ -237,14 +239,8 @@ export class Container {
     // that killAll() may have missed (e.g. sessions that were spawning during shutdown).
     const tmuxSessionManagerResult = this.get<TmuxSessionManagerCorePort>('tmuxSessionManager');
     if (tmuxSessionManagerResult.ok) {
-      const tmux = tmuxSessionManagerResult.value;
-      const sessionsResult = tmux.listSessions();
-      if (sessionsResult.ok) {
-        for (const session of sessionsResult.value) {
-          tmux.destroySession(session.name);
-          // Failure is silently ignored during dispose — don't block DB close
-        }
-      }
+      // No logger passed — failures silently ignored so DB close is not blocked
+      sweepTmuxSessions(tmuxSessionManagerResult.value);
     }
 
     // Close database if exists
